feat: uninstall agent when deleting from console

+++ internal/controllers/webserver/handlers/agents.go
@@ -276,9 +276,26 @@ func (h *Handler) AgentConfirmDelete(c echo.Context) error {
 		return RenderError(c, partials.ErrorMessage(i18n.T(c.Request().Context(), "agents.no_empty_id"), true))
 	}
 
-	err := h.Model.DeleteAgent(agentId)
-	if err != nil {
-		return h.ListAgents(c, "", err.Error(), false)
+	deleteAction := c.FormValue("agent-delete-action")
+
+	if deleteAction == "delete-and-uninstall" || deleteAction == "keep-and-uninstall" {
+		if h.NATSConnection == nil || !h.NATSConnection.IsConnected() {
+			return RenderError(c, partials.ErrorMessage(i18n.T(c.Request().Context(), "nats.not_connected"), false))
+		}
+
+		ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+		defer cancel()
+		if _, err := h.JetStream.Publish(ctx, "agent.uninstall."+agentId, nil); err != nil {
+			log.Println(err, "agent.uninstall."+agentId)
+			return RenderError(c, partials.ErrorMessage(i18n.T(c.Request().Context(), "agents.could_not_send_request_to_uninstall"), true))
+		}
+	}
+
+	if deleteAction == "delete-and-uninstall" || deleteAction == "delete-and-keep" {
+		err := h.Model.DeleteAgent(agentId)
+		if err != nil {
+			return h.ListAgents(c, "", err.Error(), false)
+		}
 	}
 
 	return h.ListAgents(c, i18n.T(c.Request().Context(), "agents.deleted"), "", true)
